// src/app/models/updateemployee.ts
export class Updateemployee{
    first_name: string;
    last_name: string;
    description: string;
    email: string;
    phone_number: string;
    age: number;
    birthday: string;
    growth: number;
    clothsize: number;
    chestsize: number;
    waistsize: number;
    thighsize: number;
    footsize: number;
    necksize: number;
    headsize: number;
    tatu: boolean;
    piercing: boolean;
    appearancetype :number;
    haircolor: number;
    city: number;
    eyecolor: number;
    gender: number;
    country: number;
    employmenttype: number;
}